Change the API for the FlowStatistics component. Instead of always passing a pair of measurement objects and a metric to display along with a value display callback to override the display value, we will allow two separate paths for passing input. When the measurment objects are passed we will display the metric value without modification and if the value needs to be modified prior to the display, then the client has to do it entirely within the calling context and pass in currentValue and previousValue instead.
This is a cleaner contract, and avoids situations like the one  in this bug. In this case we were doing trending based on the raw flow metric value, but we were display based on percentages using the display override. Hence the nonsensical result like all percentages increasing together (the underlying values were increasing together, but that does not mean the percentages can).

Along the way, we also cleaned up the input prop for measurement, renaming currentCycleMetrics -> currentMeasurement and previousCycleMetrics -> previousMeasurement. This necessitates changing all the calling sites to use the new prop, hence the need to change so many files in the commit that follows this.

[https://urjuna.atlassian.net/browse/PO-303]