Adding Support for Prettier.
Using Prettier Configuration files to set formatting options is the
recommended approach.

The Editor settings are meant to be used as a fallback and are
generally intended only for use on non-project files. 
It is recommended that you always include a prettier configuration file
in your project specifying all settings for your project.
This will ensure that no matter how you run prettier - from
the extension, from the CLI, or from another IDE with Prettier,
the same settings will get applied.