Updates the components  separating responsibilities between widget, view, chart and reducer appropriately. The code has relatively few edge cases to handle now and all are in the view as is appropriate.
Here are the main changes:

Separation of responsibilities:

1. The widget extracts the GraphQL query and hands an array of workItemSources to the view.
2. The view is responsible for managing the selection of a single work item source from this array and maintaining the updated state of the edits to the state mapping of this work item source. It delegates to the chart for displaying the selected work item source.
3. The chart is responsible for displaying the state mapping of a single work item source which is provided to it (from anywhere) (in this case it happens to be the one passed to it by the view.

So this changes the contracts for the chart so that it takes a single work item source, and the contract for the view so that it takes an array of work item sources.. none of the components need to know anything else about each other and they are pretty well independent of each other, which is the way it should be.

Additional points to note:

workItemStateTypeMapView.js:

1. Line 25 handles the empty array case on state initialization.
2. Line 139 passes the currentlySelected work item source to the chart, also needs to handle the empty array case.

Both are fairly logical places where such a check should happen because the chart expects a workItemSource which may be null.

workItemStateTypeMapChart.js:

1. Since the chart works on an array representing the mappings of single work item source, line 78 handles its initialization checking for a null workWorkSource if needed. Again, looks natural, no special cases.

workItemReducer.js:

1. takes an array of workItemSources and the selected index and updates the state based on these two and the action from the dispatch.

again - no special cases here.. every action handles what it needs based on the these parameters. The reducer returns the passed in state if the array is empty. Again a reasonable thing for this reducer to do.

All tests verified to pass.

[PO-411]