Rename getCategories in projects/shared/helpers/utils to getHistogramCategories.
Update this method to take selectedMetric and metricsMeta as args so that we
can show different units of measure for the selected metric - the use case here is
days vs dev days for effort.

Note: This is only a partial fix since this method is also used to show the
categories for the filter in the detail table. I did not refactor that table code yet and am falling back to showing days for all metrics in that filter selection.

This is a todo item that needs to be fixed, but it is not super-critical and we can address it along with the other table UX changes.

[PP-37]