[PP-109] Move the getActualMetric method to metricsMeta.js and rename it to getSelectedMetricDisplayName. We will use this method to try and centralize the logic for mapping metric keys to metric display names. It does not still quite feel right yet, since the space of inputs being mapped is not quite the same here across all the values being mapped in the method. But at least it is a marginal improvement from spreading this all across the components using this and it will put this is in the same file as metricsMeta.js definitions.
We will use the getSelectedMetricDisplayName method to map "latency" to "delivery" for closed work items. We needed to do this because we
were originally just mapping it directly in the metrics meta since the only use case for showing this was for closed items, and if we want to show latency and delivery separately for closed in the valueStreamPhaseDetailView.js, then we will need to handle these two cases separately. This prepare us for this use case.