[PP-212] The changes here are the heart of the issue why detail navigation was not working. When a topic dashboard is nested below another topic dashboard, the path to the topic  needs to be explicity added to context route tree when buildRoutes is called. The way to wire this up is to use the "topic" property in the Topic route and not the directly wire up the component at this level. This is the change on line 12 and 18.
There is a bit of awkwardness is that now we have to interpolate the topic properties both at the parent level
and since the subnav looks for those properties at this level - I think we can clean this up in the sidebar-subnav
component since we can assume that if subnav is true there will be a property on the route which is called topic and
the display and icon properties of this nested topic can be used to display the subnav. I'll leave this change to you, let me know if this breaks due to some other assumption that I may have missed.

If we make this change then
code and topic structure for sub topics should look just like those of regular topics with the only remaining assumption being that we are assuming subnavs can be applied only at the top level of a topic. ie we cannot have subnavs within subnavs without a more sophisticated sidebar component. I dont think we need to consider this case.