PP-540: Change the contract of quadrantSummaryPanel.js to take out state types from the api and just display the quadrant summary based on what was passed to it in the work items. The expected contract is now: - work items have been passed through the getWorkItemDurations method and have the latency etc augmented to it. - If you want to filter by state types, the caller should do it.
This seems like a more generic contract for quadrant summary panel, especially now we are doing a lot of other types of filtering outside the widget.

This change was also motivated by the fact that we need the same filtered list passed to both the motion effiency calculation on the motionEfficiencyQuadrantSummaryCard.js and so it did not seem to make sense to calculate this in two places every time. (This is the changes on line 9-10 of motionEfficiencyQuadrantSummaryCard.js)

So, now wherever we use quadrantSummaryPanel.js we need to filter it by stateTypes if this is a context where stateTypes are being passed to it.